change categorise page to products page change landing page name to home page fix some bugs on products page
